Refactor name of packages to project name FlashCLI
