+++ src/test/java/FlashCLI/FlashCLITest.java
@@ -1,10 +1,10 @@
-package seedu.duke;
+package FlashCLI;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class FlashCLITest {
     @Test
     public void sampleTest() {
         assertTrue(true);
